refactor: BottomSheet 컴포넌트 interface 검토 및 리팩토링 적용
