+++ src/pages/SignUp.tsx
@@ -6,13 +6,13 @@ import { ReactComponent as Close } from '../assets/svg/close_24.svg';
 import useCheckbox from '../components/@common/@hooks/useCheckbox.ts';
 import useInput from '../components/@common/@hooks/useInput.ts';
 import useRadio from '../components/@common/@hooks/useRadio.ts';
+import BottomSheet from '../components/@common/BottomSheet/BottomSheet.tsx';
 import Button from '../components/@common/Button/Button.tsx';
 import CheckboxContainer from '../components/@common/CheckboxContainer/CheckboxContainer.tsx';
 import Dropdown from '../components/@common/Dropdown/Dropdown.tsx';
 import Field from '../components/@common/Field/Field.tsx';
 import RadioContainer from '../components/@common/RadioContainer/RadioContainer.tsx';
 import Text from '../components/@common/Text/Text.tsx';
-import BottomSheet from '../components/BottomSheet/BottomSheet.tsx';
 import CheckBoxModal from '../components/BottomSheet/CheckBox.tsx';
 import IdeaCard from '../components/Card/IdeaCard.tsx';
 import PopCard from '../components/Card/PopCard.tsx';
